Collect dropped files for sending to the host
Close #31

+++ analyzer/darwin/lib/dtrace/apicalls.d
@@ -79,11 +79,11 @@ pid$target::execve:entry
     this->retval = 0;
     this->timestamp_ms = walltimestamp/1000000;
 
-    printf("{\"api\":\"%s\", \"args\":[\"%S\", %llu, %llu], \"retval\":%d, \"timestamp\":%ld, \"pid\":%d, \"ppid\":%d, \"tid\":%d, \"errno\":%d}\n",
+    printf("{\"api\":\"%s\", \"args\":[\"%S\", %llu, %llu], \"retval\":%d, \"timestamp\":%lld, \"pid\":%d, \"ppid\":%d, \"tid\":%d, \"errno\":%d}\n",
         probefunc,
         copyinstr(arg0), (unsigned long long)arg1, (unsigned long long)arg2,
         (int)this->retval,
-        this->timestamp_ms, pid, ppid, tid, errno);
+        (int64_t)this->timestamp_ms, pid, ppid, tid, errno);
 }
 
 #endif /* not SUDO */
